requirements.txt
openai
langchain
feedparser
requests
beautifulsoup4
python-dotenv